pyproject.toml
[tool.poetry]
name = "oblique-strategy-games"
version = "0.1.0"
description = ""
authors = ["reayd-falmouth <dr323090@falmouth.ac.uk>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
pygame-gui = "^0.6.13"
moderngl = "^5.12.0"
pygame-ce = "^2.5.3"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pillow = "^11.1.0"
pygbag = "^0.9.2"
pyinstaller = "^6.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
